update the integration test stubs
Was expecting to go to Home, but The Book says got to user profile, which is
easier to test for anyways.  I also moved the login steps file into it's
rightful home, but since it was working (well, it found the step definitions
at least), I'm condiering ditching the step_definitions directory. Thats how
I've always done it - with a separate directory - but that might be
needlessly complex, as the file name and extension semantically identify
them as step definition files anyways. I'll add that to my list of minor
tasks to do when the important stuff is done.