Add the actual oauth code
This code was blatantly ripped from the (example
code)[https://github.com/500px/api-documentation/blob/master/examples/Ruby/xauth.rb]

I'm going to actually store the access token and secret, and build up a
new Oauth consumer object for every request.  Obviously, asking users
for their 500px credentials in my app instead of doing 3-legged
authentication is sketchy and terrible, but I spent a good day and a
half trying to do it 'right', without much in the way of success.  I
would *never ever* do this in a real app, but it gets the authentication
stuff out of the way/