Add the like button (link)
just wires up the action for now. What will happen is this makes a
PATCH (update) request to the app, for the photo in question, which
turns around and sends a positive vote request to the 500px api.

Later on, I'll keep track of which photos a user has liked, and only
show this like button if they haven't liked a photo. I'll either store
this liked list in my app, or leverage the 500px api.  The latter is
probably more 'correct', but I don't know if I can pull it off as
quickly.  using the 500px api is my preference, but time is finite, and
I'd like to finish this before the sun expands into a red giant.