Add seed reset after random tests
I had to set the seed to a consistent value to test the random_string
function, but I forgot to set it back.  So afterwards, when I used
random_string to make usernames and passwords, it would keep returning
the same strong, erroring out the signup page.  Now it gets reset to a
timestamp, so collisions should be virtually impossible now.