add voting feedback and test
Originally, my plan here was to have the test vote for a photo, check
that the vote count went up, and then remove the vote, so as to not
pollute the database.  Unfortunately, I couldn't find a way to roll back
the vote.  I could make a sort of proxy-mock server that passes GET
requests though, but intercepts POST requests, but that would have been
too complex to do quickly.

Instead, I added this feedback (which is good to have anyways) and look
for that.  I made the 'already_voted' feedback show as success because
to the user, that operation should be idempotent. It also makes the test
easier.

I also fixed a typo in my test credentials. oops!